fix: use sha instead of branch
My attempt to solve https://github.com/apiaryio/black-belt/issues/111. Not tested, this is pure shooting into dark. My idea is we cannot create the deployment using the branch as a ref, because we delete the branch after merging. Instead, we could probably use the sha. Let's see. It's already broken, so this change cannot make it more broken. Hopefully.